fix(frontend): convert folder route to layout with Outlet for child routes
The $folderName.tsx route was rendering the FolderDetailPage directly
without an <Outlet />, which prevented child routes like
$folderName/templates.tsx from ever rendering. This caused E2E tests
for folder templates to fail because navigating to
/orgs/$orgName/folders/$folderName/templates showed the folder detail
page instead of the templates page.

Split into:
- $folderName.tsx: layout route that renders <Outlet />
- $folderName/index.tsx: folder detail page (moved content)
